Insert your own image and make it animated.

Insert your own image and make it animated.

cia(src, height = NULL, width = NULL, size = 1, position = c(0, 0, 0,
  0), grow = 0, animate = "flash", speed = "normal", anitype = "repeat",
  rotate = 0, flip = "none", rtext = "", ltext = "", border = FALSE,
  bgcolor = NULL, bgcolour = bgcolor, iother = NULL, sother = NULL,
  dother = NULL)

Arguments

src
A url or path of image/gif to animate.

height
Height of the image that can be specified in percentage or pixels. This cannot be defined if size or grow is defined.

width
Width of the image that can be specified in percentage or pixels. This cannot be defined if size or grow is defined.

size
Size of the icon relative to font size. Options are lg (133 percent), xs (75 percent), sm (87.5 percent) or a positive number.

position
A numerical vector of length 4 specifying the value to move up, down, left and right.

grow
Numerical value to grow (or shrink for negative values) without changing or moving the container.

animate
'bounce', 'burst', 'falling', 'flash', 'float', 'horizontal', 'passing', 'passing-reverse', 'pulse', 'ring', 'shake', 'spin', 'tada', 'vertical', 'wrench', or use FALSE not to animate.

speed
'normal', 'fast', or 'slow'

anitype
'repeat', 'hover', or 'parent-hover'

rotate
Numerical value for degree of rotation.

flip
'none', 'horizontal', 'vertical'.

rtext
The text to be inserted on the right side.

ltext
The text to be inserted on the left side.

border
If TRUE, draws a border around the icon.

bgcolor, bgcolour
Colour to be given to the background

sother
Character vector of other parameters directly added to the style classes.

References

Font awesome animation